Revert "[PM-10706] Moving webauthn-specific files from Vault ownership to autofill and fixing issue with excluded domains showing passkeys within inline menu (#10436)"
This reverts commit 48cb6fbec41e8bf84548bbe4f334eacbdbde0fc6.
