+++ apps/browser/src/_locales/en/messages.json
@@ -3946,11 +3946,11 @@
   },
   "passkeys": {
     "message": "Passkeys",
-    "description": "A section header for a list of passkeys."
+    "description": "A section header for a list of passkeys. Used in the inline menu list."
   },
   "passwords": {
     "message": "Passwords",
-    "description": "A section header for a list of passwords."
+    "description": "A section header for a list of passwords. Used in the inline menu list."
   },
   "logInWithPasskeyAriaLabel": {
     "message": "Log in with passkey",
