+++ apps/browser/src/autofill/background/overlay.background.ts
@@ -7,7 +7,6 @@ import {
   switchMap,
   debounceTime,
 } from "rxjs";
-import { parse } from "tldts";
 
 import { AuthService } from "@bitwarden/common/auth/abstractions/auth.service";
 import { AuthenticationStatus } from "@bitwarden/common/auth/enums/authentication-status";
@@ -18,7 +17,6 @@ import {
 import { AutofillSettingsServiceAbstraction } from "@bitwarden/common/autofill/services/autofill-settings.service";
 import { DomainSettingsService } from "@bitwarden/common/autofill/services/domain-settings.service";
 import { InlineMenuVisibilitySetting } from "@bitwarden/common/autofill/types";
-import { NeverDomains } from "@bitwarden/common/models/domain/domain-service";
 import { EnvironmentService } from "@bitwarden/common/platform/abstractions/environment.service";
 import { Fido2ClientService } from "@bitwarden/common/platform/abstractions/fido2/fido2-client.service.abstraction";
 import { I18nService } from "@bitwarden/common/platform/abstractions/i18n.service";
@@ -366,10 +364,7 @@ export class OverlayBackground implements OverlayBackgroundInterface {
         true,
       );
     } else {
-      inlineMenuCipherData = await this.buildInlineMenuCiphers(
-        inlineMenuCiphersArray,
-        showFavicons,
-      );
+      inlineMenuCipherData = this.buildInlineMenuCiphers(inlineMenuCiphersArray, showFavicons);
     }
 
     this.currentInlineMenuCiphersCount = inlineMenuCipherData.length;
@@ -437,17 +432,12 @@ export class OverlayBackground implements OverlayBackgroundInterface {
    * @param inlineMenuCiphersArray - Array of inline menu ciphers
    * @param showFavicons - Identifies whether favicons should be shown
    */
-  private async buildInlineMenuCiphers(
+  private buildInlineMenuCiphers(
     inlineMenuCiphersArray: [string, CipherView][],
     showFavicons: boolean,
   ) {
     const inlineMenuCipherData: InlineMenuCipherData[] = [];
     const passkeyCipherData: InlineMenuCipherData[] = [];
-    const domainExclusions = await this.getExcludedDomains();
-    let domainExclusionsSet: Set<string> | null = null;
-    if (domainExclusions) {
-      domainExclusionsSet = new Set(Object.keys(await this.getExcludedDomains()));
-    }
 
     for (let cipherIndex = 0; cipherIndex < inlineMenuCiphersArray.length; cipherIndex++) {
       const [inlineMenuCipherId, cipher] = inlineMenuCiphersArray[cipherIndex];
@@ -455,7 +445,7 @@ export class OverlayBackground implements OverlayBackgroundInterface {
         continue;
       }
 
-      if (this.showCipherAsPasskey(cipher, domainExclusionsSet)) {
+      if (this.showCipherAsPasskey(cipher)) {
         passkeyCipherData.push(
           this.buildCipherData({
             inlineMenuCipherId,
@@ -482,28 +472,13 @@ export class OverlayBackground implements OverlayBackgroundInterface {
    * Identifies whether we should show the cipher as a passkey in the inline menu list.
    *
    * @param cipher - The cipher to check
-   * @param domainExclusions - The domain exclusions to check against
    */
-  private showCipherAsPasskey(cipher: CipherView, domainExclusions: Set<string> | null): boolean {
-    if (cipher.type !== CipherType.Login) {
-      return false;
-    }
-
-    const fido2Credentials = cipher.login.fido2Credentials;
-    if (!fido2Credentials?.length) {
-      return false;
-    }
-
-    const credentialId = fido2Credentials[0].credentialId;
-    const rpId = fido2Credentials[0].rpId;
-    const parsedRpId = parse(rpId, { allowPrivateDomains: true });
-    if (domainExclusions?.has(parsedRpId.domain)) {
-      return false;
-    }
-
+  private showCipherAsPasskey(cipher: CipherView): boolean {
     return (
-      this.inlineMenuFido2Credentials.size === 0 ||
-      this.inlineMenuFido2Credentials.has(credentialId)
+      cipher.type === CipherType.Login &&
+      cipher.login.fido2Credentials?.length > 0 &&
+      (this.inlineMenuFido2Credentials.size === 0 ||
+        this.inlineMenuFido2Credentials.has(cipher.login.fido2Credentials[0].credentialId))
     );
   }
 
@@ -622,17 +597,9 @@ export class OverlayBackground implements OverlayBackgroundInterface {
    * @param credentials - The FIDO2 credentials to store
    */
   private storeInlineMenuFido2Credentials(credentials: Fido2CredentialView[]) {
-    credentials.forEach(
-      (credential) =>
-        credential?.credentialId && this.inlineMenuFido2Credentials.add(credential.credentialId),
-    );
-  }
-
-  /**
-   * Gets the neverDomains setting from the domain settings service.
-   */
-  async getExcludedDomains(): Promise<NeverDomains> {
-    return await firstValueFrom(this.domainSettingsService.neverDomains$);
+    credentials
+      .map((credential) => credential.credentialId)
+      .forEach((credentialId) => this.inlineMenuFido2Credentials.add(credentialId));
   }
 
   /**
