+++ apps/browser/src/autofill/fido2/content/fido2-page-script.webauthn-supported.spec.ts
@@ -5,7 +5,7 @@ import {
   createCredentialRequestOptionsMock,
   setupMockedWebAuthnSupport,
 } from "../../../autofill/spec/fido2-testing-utils";
-import { WebauthnUtils } from "../utils/webauthn-utils";
+import { WebauthnUtils } from "../../../vault/fido2/webauthn-utils";
 
 import { MessageType } from "./messaging/message";
 import { Messenger } from "./messaging/messenger";
@@ -41,7 +41,7 @@ jest.mock("./messaging/messenger", () => {
     },
   };
 });
-jest.mock("../utils/webauthn-utils");
+jest.mock("../../../vault/fido2/webauthn-utils");
 
 describe("Fido2 page script with native WebAuthn support", () => {
   (jest.spyOn(globalThis, "document", "get") as jest.Mock).mockImplementation(
